Accept floating-point coordinates in DrawRectangle/DrawLine/DrawSprite
Changes DrawRectangle, DrawLine, and DrawSprite to take float
coordinates/dimensions instead of int, since gameplay positions are
naturally float (updated via delta time) and every continuous-movement
caller was casting at the draw call site. Backend implementation now
calls raylib's float-native DrawRectangleRec/DrawLineV/DrawTextureV
directly instead of the int-based variants, so no rounding/truncation
policy is introduced by the engine itself.

Existing integer callers are untouched and keep working via ordinary
int->float conversion.

Removes now-obsolete static_cast<int> calls from 03_input,
08_geometry, prototype_01_continuous, and prototype_02_projectiles.
Adds 09_float_rendering, demonstrating a continuously bouncing
rectangle drawn straight from a float position with no casts anywhere
in the example.

+++ engine/include/engine/Engine.hpp
@@ -92,8 +92,8 @@ public:
     void Clear(Color color);
     void DrawText(const char* text, int x, int y, int fontSize, Color color);
     void DrawText(const std::string& text, int x, int y, int fontSize, Color color);
-    void DrawRectangle(int x, int y, int width, int height, Color color);
-    void DrawLine(int x1, int y1, int x2, int y2, Color color);
+    void DrawRectangle(float x, float y, float width, float height, Color color);
+    void DrawLine(float x1, float y1, float x2, float y2, Color color);
 
     // Loads a texture and hands back a handle to it. The engine owns the
     // texture from this point on; there is no explicit unload — all loaded
@@ -103,7 +103,7 @@ public:
     TextureHandle LoadTexture(const char* filePath);
     int TextureWidth(TextureHandle texture) const;
     int TextureHeight(TextureHandle texture) const;
-    void DrawSprite(TextureHandle texture, int x, int y);
+    void DrawSprite(TextureHandle texture, float x, float y);
 
     // Diagnostic: how many distinct textures are currently loaded. Useful
     // for verifying that repeated LoadTexture calls are being deduplicated;
